+++ jgclark.RepeatExtensions/src/repeatExtensions.js
@@ -2,7 +2,7 @@
 //-----------------------------------------------------------------------
 // Repeat Extensions plugin for NotePlan
 // Jonathan Clark
-// last updated 14.3.2023 for v0.5.2
+// last updated 7.7.2023 for v0.6.0
 //-----------------------------------------------------------------------
 
 import pluginJson from "../plugin.json"
@@ -12,6 +12,10 @@ import {
   getISOWeekString,
   isDailyNote,
   isWeeklyNote,
+  isMonthlyNote,
+  isQuarterlyNote,
+  isYearlyNote,
+  RE_ANY_DUE_DATE_TYPE,
   RE_DATE, // find dates of form YYYY-MM-DD
   RE_DATE_INTERVAL,
   RE_DATE_TIME,
@@ -19,6 +23,9 @@ import {
   RE_DONE_DATE_TIME_CAPTURES,
   RE_SCHEDULED_DAILY_NOTE_LINK,
   RE_SCHEDULED_WEEK_NOTE_LINK,
+  RE_SCHEDULED_MONTH_NOTE_LINK,
+  RE_SCHEDULED_QUARTERLY_NOTE_LINK,
+  RE_SCHEDULED_YEARLY_NOTE_LINK,
   RE_TIME, // find '12:23' with optional '[ ][AM|PM|am|pm]'
   unhyphenateString,
 } from '@helpers/dateTime'
@@ -109,30 +116,19 @@ export async function onEditorWillSave(): Promise<void> {
  * Valid intervals are [0-9][bdwmqy].
  * To work it relies on finding @done(YYYY-MM-DD HH:MM) tags that haven't yet been shortened to @done(YYYY-MM-DD).
  * It includes cancelled tasks as well; to remove a repeat entirely, remove the @repeat tag from the task in NotePlan.
- * Note: The new repeat date is by default scheduled to a day (>YYYY-MM-DD).
- * But if the scheduled date is a week date (YYYY-Wnn), or the repeat is in a weekly note, then the new repeat date will be a scheduled week link (>YYYY-Wnn).
+ * Note: The new repeat date is by default scheduled to a day (>YYYY-MM-DD). But if the scheduled date is a week date (YYYY-Wnn), or the repeat is in a weekly note, then the new repeat date will be a scheduled week link (>YYYY-Wnn).
  * Note: Runs on the currently open note (using Editor.* funcs)
  * Note: Could add a 'Newer' mode of operation according to #351.
  * TEST: fails to appendTodo to note with same stem?
  * @author @jgclark
- * @param {CoreNoteFields} noteIn?
+ * @param {boolean} runSilently?
  */
-export async function repeats(showMessages: boolean = true): Promise<void> {
+export async function repeats(runSilently: boolean = false): Promise<void> {
   try {
-    // Get passed note details, or fall back to Editor
-    // Note: v0.5.2 changed this to run from 'Editor.note' only
-    // let note: CoreNoteFields
-    // let showMessages: boolean
-    // if (noteIn) {
-    //   note = noteIn
-      // let showMessages = false
-    // } else {
-      if (!Editor.note) {
-        throw new Error(`repeats: Couldn't get Editor.note to process`)
-      }
-    //   note = Editor.note
-    //   showMessages = true
-    // }
+    // Get passed note details, or fall back to Editor. (Note: v0.5.2 changed this to run from 'Editor.note' only)
+    if (!Editor.note) {
+      throw new Error(`repeats: Couldn't get Editor.note to process`)
+    }
     const { paragraphs, title, type, filename, note } = Editor
     if (note === null || paragraphs === null) {
       // No note open, or no paragraphs (perhaps empty note), so don't do anything.
@@ -149,8 +145,8 @@ export async function repeats(showMessages: boolean = true): Promise<void> {
     // work out where ## Done or ## Cancelled sections start, if present
     const endOfActive = findEndOfActivePartOfNote(note)
 
-    logDebug(pluginJson, `Starting for '${filename}' for ${endOfActive} active lines, showMessages: ${String(showMessages)}`)
-    logAllEnvironmentSettings()
+    logDebug(pluginJson, `Starting for '${filename}' for ${endOfActive} active lines`)
+    // logAllEnvironmentSettings()
 
     let repeatCount = 0
     let line = ''
@@ -184,24 +180,34 @@ export async function repeats(showMessages: boolean = true): Promise<void> {
         logDebug('repeats', `- updated Para ${p.lineIndex} -> <${updatedLine}>`)
 
         // Test if this is one of my special extended repeats
+
+        // TODO: Split most into a separate function so it can be tested externally
+
         if (updatedLine.match(RE_EXTENDED_REPEAT)) {
           repeatCount++
-          let newRepeatDateStr = ''
-          let newRepeatDate: Date
-          let outputLine = ''
-
           // get repeat to apply
           reReturnArray = updatedLine.match(RE_EXTENDED_REPEAT_CAPTURE) ?? []
           let dateIntervalString: string = (reReturnArray.length > 0) ? reReturnArray[1] : ''
-          logDebug('repeats', `- Found EXTENDED @repeat syntax: '${dateIntervalString}'`)
+          logDebug('repeats', `- Found extended @repeat syntax: '${dateIntervalString}'`)
 
-          // decide style of new date: daily or weekly link
+          // decide style of new date: daily / weekly / monthly / etc.link
           let timeframe = 'day'
-          if (updatedLine.match(RE_SCHEDULED_WEEK_NOTE_LINK) || isWeeklyNote(note)) {
+          // TEST: should this be now without test "or isWeeklyNote"?
+          if (updatedLine.match(RE_SCHEDULED_WEEK_NOTE_LINK) /** || isWeeklyNote(note)*/) {
             timeframe = 'week'
+          } else if (updatedLine.match(RE_SCHEDULED_MONTH_NOTE_LINK)) {
+            timeframe = 'month'
+          } else if (updatedLine.match(RE_SCHEDULED_QUARTERLY_NOTE_LINK)) {
+            timeframe = 'quarter'
+          } else if (updatedLine.match(RE_SCHEDULED_YEARLY_NOTE_LINK) /** || isYearlyNote(note)*/) {
+            timeframe = 'year'
           }
-          // TODO: use this in >day case in weekly note
-          logDebug('repeats', `- timeframe: ${timeframe} from ${String(updatedLine.match(RE_SCHEDULED_WEEK_NOTE_LINK))} / ${String(isWeeklyNote(note))}`)
+          // TODO: use this in >day case in weekly note. TEST: now done?
+          logDebug('repeats', `- timeframe: ${timeframe}`)
+
+          let newRepeatDateStr = ''
+          let newRepeatDate: Date
+          let outputLine = ''
 
           if (dateIntervalString[0].startsWith('+')) {
             // New repeat date = completed date (of form YYYY-MM-DD) + interval
@@ -209,28 +215,32 @@ export async function repeats(showMessages: boolean = true): Promise<void> {
               1,
               dateIntervalString.length,
             )
+            newRepeatDate = calcOffsetDate(completedDate, dateIntervalString) ?? new moment().startOf('day')
             newRepeatDateStr = calcOffsetDateStr(completedDate, dateIntervalString)
-            newRepeatDate = calcOffsetDate(completedDate, dateIntervalString) ?? new Date()
             logDebug('repeats', `- Adding from completed date -> ${newRepeatDateStr}, ${String(newRepeatDate)}`)
             // Remove any >date
-            // Note: ' >'+RE_DUE_DATE, but can't get regex to work with variables like this
-            updatedLine = updatedLine.replace(/\s+>\d{4}-[01]\d{1}-\d{2}/, '')
-            // TODO: or week version?
+            updatedLine = updatedLine.replace(RE_ANY_DUE_DATE_TYPE, '')
             logDebug('repeats', `- updatedLine: ${updatedLine}`)
 
           } else {
             // New repeat date = due date + interval
-            // look for the due date (>YYYY-MM-DD)
+            // look for the due date (>YYYY-MM-DD) or other calendar types
             let dueDate = ''
             const dueDateArray = RE_SCHEDULED_DAILY_NOTE_LINK.test(updatedLine)
               ? updatedLine.match(RE_SCHEDULED_DAILY_NOTE_LINK)
               : RE_SCHEDULED_WEEK_NOTE_LINK.test(updatedLine)
                 ? updatedLine.match(RE_SCHEDULED_WEEK_NOTE_LINK)
+                : RE_SCHEDULED_MONTH_NOTE_LINK.test(updatedLine)
+                  ? updatedLine.match(RE_SCHEDULED_MONTH_NOTE_LINK)
+                  : RE_SCHEDULED_QUARTERLY_NOTE_LINK.test(updatedLine)
+                    ? updatedLine.match(RE_SCHEDULED_QUARTERLY_NOTE_LINK)
+                    : RE_SCHEDULED_YEARLY_NOTE_LINK.test(updatedLine)
+                      ? updatedLine.match(RE_SCHEDULED_YEARLY_NOTE_LINK)
                 : []
             logDebug('repeats', `- dueDateArray: ${String(dueDateArray)}`)
             if (dueDateArray && dueDateArray[0] != null) {
               dueDate = dueDateArray[0].split('>')[1]
-              // logDebug('repeats', `  due date match => ${dueDate}`)
+              logDebug('repeats', `  due date match = ${dueDate}`)
               // need to remove the old due date
               updatedLine = updatedLine.replace(` >${dueDate}`, '')
             } else {
@@ -239,8 +249,8 @@ export async function repeats(showMessages: boolean = true): Promise<void> {
               logDebug('repeats', `- no match => use completed date ${dueDate}`)
             }
             newRepeatDateStr = calcOffsetDateStr(dueDate, dateIntervalString)
-            newRepeatDate = calcOffsetDate(dueDate, dateIntervalString) ?? new Date()
-            logDebug('repeats', `- adding from due date -> ${newRepeatDateStr}, ${String(newRepeatDate)}`)
+            // newRepeatDate = calcOffsetDate(dueDate, dateIntervalString) ?? new moment().startOf('day')
+            logDebug('repeats', `- adding from due date -> ${newRepeatDateStr}`)
           }
 
           outputLine = updatedLine.replace(/@done\(.*\)/, '').trim()
@@ -272,9 +282,9 @@ export async function repeats(showMessages: boolean = true): Promise<void> {
             // }
             logInfo(pluginJson, `- repeat -> ${newRepeatDateStr}`)
 
-            // Grr: can't use this next one, as it doesn't support undefined notes
-            // let futureNote = await DataStore.calendarNoteByDateString(newRepeatDateStr)
-            let futureNote = await DataStore.calendarNoteByDate(newRepeatDate, timeframe)
+            // Get future note (if it exists)
+            let futureNote = await DataStore.calendarNoteByDateString(newRepeatDateStr)
+            // let futureNote = await DataStore.calendarNoteByDate(newRepeatDate, timeframe)
             if (futureNote != null) {
               // Add todo to future note
               logDebug(pluginJson, futureNote.filename)
@@ -294,7 +304,7 @@ export async function repeats(showMessages: boolean = true): Promise<void> {
     }
     if (repeatCount === 0) {
       logWarn('repeats', 'No suitable completed repeats found')
-      if (showMessages) {
+      if (!runSilently) {
         await showMessage('No suitable completed repeats found', 'OK', 'Repeat Extensions')
       }
     }